refactor: flow control, move plugin resolution into the config loader
