+++ pkg/epp/flowcontrol/benchmark/benchmark.go
@@ -74,7 +74,6 @@ import (
 	"github.com/llm-d/llm-d-router/pkg/epp/flowcontrol/registry"
 	fwkdl "github.com/llm-d/llm-d-router/pkg/epp/framework/interface/datalayer"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/interface/flowcontrol"
-	"github.com/llm-d/llm-d-router/pkg/epp/framework/interface/plugin"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/interface/scheduling"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/plugins/flowcontrol/fairness/globalstrict"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/plugins/flowcontrol/ordering/fcfs"
@@ -190,7 +189,7 @@ func (r *benchRequest) ReceivedTimestamp() time.Time                   { return 
 func setupRegistry(
 	ctx context.Context,
 	b *testing.B,
-	handle plugin.Handle,
+	defaults registry.PriorityBandPolicyDefaults,
 	p priorityLevels,
 ) contracts.FlowRegistry {
 	b.Helper()
@@ -201,7 +200,7 @@ func setupRegistry(
 
 	for i := 0; i < int(p); i++ {
 		band, err := registry.NewPriorityBandConfig(
-			handle, i,
+			i, defaults,
 			registry.WithBandMaxBytes(10_000_000_000), // Prevent capacity-based rejections.
 		)
 		if err != nil {
@@ -210,7 +209,7 @@ func setupRegistry(
 		cfgOpts = append(cfgOpts, registry.WithPriorityBand(band))
 	}
 
-	regCfg, err := registry.NewConfig(handle, cfgOpts...)
+	regCfg, err := registry.NewConfig(defaults, cfgOpts...)
 	if err != nil {
 		b.Fatalf("Failed to create registry config: %v", err)
 	}
@@ -245,7 +244,12 @@ func setupBenchmarkHarness(
 	}
 	handle.AddPlugin(registry.DefaultOrderingPolicyRef, oPolicy)
 
-	reg := setupRegistry(ctx, b, handle, p)
+	defaults := registry.PriorityBandPolicyDefaults{
+		OrderingPolicy: oPolicy.(flowcontrol.OrderingPolicy),
+		FairnessPolicy: fPolicy.(flowcontrol.FairnessPolicy),
+	}
+
+	reg := setupRegistry(ctx, b, defaults, p)
 
 	detector := customDetector
 	if detector == nil {
