+++ pkg/epp/flowcontrol/registry/config.go
@@ -22,13 +22,9 @@ import (
 	"slices"
 	"time"
 
-	"k8s.io/apimachinery/pkg/api/resource"
-
-	configapi "github.com/llm-d/llm-d-router/apix/config/v1alpha1"
 	"github.com/llm-d/llm-d-router/pkg/epp/flowcontrol/contracts"
 	"github.com/llm-d/llm-d-router/pkg/epp/flowcontrol/framework/plugins/queue"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/interface/flowcontrol"
-	"github.com/llm-d/llm-d-router/pkg/epp/framework/interface/plugin"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/plugins/flowcontrol/fairness/globalstrict"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/plugins/flowcontrol/ordering/fcfs"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/plugins/flowcontrol/usagelimits"
@@ -104,6 +100,14 @@ func (r *runtimeCapabilityChecker) CheckCompatibility(p flowcontrol.OrderingPoli
 }
 
 // --- Configuration ---
+
+// PriorityBandPolicyDefaults carries pre-resolved default policy instances.
+// It is populated by the config loader (the single boundary for plugin resolution)
+// and passed into registry constructors so they never need to access the plugin Handle.
+type PriorityBandPolicyDefaults struct {
+	OrderingPolicy flowcontrol.OrderingPolicy
+	FairnessPolicy flowcontrol.FairnessPolicy
+}
 
 // Config holds the master configuration for the entire FlowRegistry.
 // It serves as the top-level blueprint, defining global settings and the templates for its priority bands.
@@ -308,56 +312,26 @@ func withCapabilityChecker(checker capabilityChecker) ConfigOption {
 // PriorityBandConfigOption defines a functional option for configuring a single PriorityBandConfig.
 type PriorityBandConfigOption func(*PriorityBandConfig) error
 
-// WithOrderingPolicy sets the name/reference of the inter-flow fairness policy (e.g., "fcfs-ordering-policy").
-// TODO(kubernetes-sigs/gateway-api-inference-extension#1794): This option is primarily used by the configuration
-// loader to wire up policies instantiated from the plugin registry.
-func WithOrderingPolicy(ref string, handle plugin.Handle) PriorityBandConfigOption {
+// WithOrderingPolicy sets the ordering policy instance for this priority band.
+func WithOrderingPolicy(policy flowcontrol.OrderingPolicy) PriorityBandConfigOption {
 	return func(p *PriorityBandConfig) error {
-		policy, err := orderingPolicy(ref, handle)
-		if err != nil {
-			return err
+		if policy == nil {
+			return errors.New("OrderingPolicy cannot be nil")
 		}
 		p.OrderingPolicy = policy
 		return nil
 	}
 }
 
-func orderingPolicy(ref string, handle plugin.Handle) (flowcontrol.OrderingPolicy, error) {
-	v := handle.Plugin(ref)
-	if v == nil {
-		return nil, fmt.Errorf("no ordering policy registered for name %q", ref)
-	}
-	policy, ok := v.(flowcontrol.OrderingPolicy)
-	if !ok {
-		return nil, fmt.Errorf("plugin %q is not a flowcontrol.OrderingPolicy (type: %T)", ref, v)
-	}
-	return policy, nil
-}
-
-// WithFairnessPolicy sets the name/reference of the inter-flow fairness policy (e.g., "round-robin-fairness-policy").
-// TODO(kubernetes-sigs/gateway-api-inference-extension#1794): This option is primarily used by the configuration
-// loader to wire up policies instantiated from the plugin registry.
-func WithFairnessPolicy(ref string, handle plugin.Handle) PriorityBandConfigOption {
+// WithFairnessPolicy sets the fairness policy instance for this priority band.
+func WithFairnessPolicy(policy flowcontrol.FairnessPolicy) PriorityBandConfigOption {
 	return func(p *PriorityBandConfig) error {
-		policy, err := fairnessPolicy(ref, handle)
-		if err != nil {
-			return err
+		if policy == nil {
+			return errors.New("FairnessPolicy cannot be nil")
 		}
 		p.FairnessPolicy = policy
 		return nil
 	}
-}
-
-func fairnessPolicy(ref string, handle plugin.Handle) (flowcontrol.FairnessPolicy, error) {
-	v := handle.Plugin(ref)
-	if v == nil {
-		return nil, fmt.Errorf("no fairness policy registered for name %q", ref)
-	}
-	policy, ok := v.(flowcontrol.FairnessPolicy)
-	if !ok {
-		return nil, fmt.Errorf("plugin %q is not a flowcontrol.FairnessPolicy (type: %T)", ref, v)
-	}
-	return policy, nil
 }
 
 // WithQueue sets the queue implementation (e.g., "ListQueue") for flows in this band.
@@ -387,143 +361,13 @@ func WithBandMaxRequests(maxRequests uint64) PriorityBandConfigOption {
 	}
 }
 
-// --- Constructors ---
-
-// resolveQuantity extracts and validates a resource.Quantity value.
-// Returns 0 if q is nil.
-func resolveQuantity(q *resource.Quantity, fieldName string) (uint64, error) {
-	if q == nil {
-		return 0, nil
-	}
-	v := q.Value()
-	if v < 0 {
-		return 0, fmt.Errorf("%s must be non-negative, got %d", fieldName, v)
-	}
-	return uint64(v), nil
-}
-
-// NewConfigFromAPI creates a new Config by translating the API configuration.
-func NewConfigFromAPI(apiConfig *configapi.FlowControlConfig, handle plugin.Handle) (*Config, error) {
-	if apiConfig == nil {
-		return NewConfig(handle)
-	}
-
-	opts := make([]ConfigOption, 0, len(apiConfig.PriorityBands)+3)
-
-	maxBytes, err := resolveQuantity(apiConfig.MaxBytes, "global MaxBytes")
-	if err != nil {
-		return nil, err
-	}
-	if maxBytes > 0 {
-		opts = append(opts, WithMaxBytes(maxBytes))
-	}
-
-	maxRequests, err := resolveQuantity(apiConfig.MaxRequests, "global MaxRequests")
-	if err != nil {
-		return nil, err
-	}
-	if maxRequests > 0 {
-		opts = append(opts, WithMaxRequests(maxRequests))
-	}
-
-	if apiConfig.DefaultPriorityBand != nil {
-		templateBand, err := buildDefaultPriorityBandTemplate(handle, apiConfig.DefaultPriorityBand)
-		if err != nil {
-			return nil, err
-		}
-		opts = append(opts, WithDefaultPriorityBand(templateBand))
-	}
-
-	if apiConfig.DefaultNegativePriorityBand != nil {
-		templateBand, err := buildDefaultPriorityBandTemplate(handle, apiConfig.DefaultNegativePriorityBand)
-		if err != nil {
-			return nil, err
-		}
-		opts = append(opts, WithDefaultNegativePriorityBand(templateBand))
-	}
-
-	for _, band := range apiConfig.PriorityBands {
-		pb, err := buildPriorityBand(handle, band)
-		if err != nil {
-			return nil, err
-		}
-		opts = append(opts, WithPriorityBand(pb))
-	}
-
-	return NewConfig(handle, opts...)
-}
-
-func buildDefaultPriorityBandTemplate(
-	handle plugin.Handle,
-	apiBand *configapi.PriorityBandConfig,
-) (*PriorityBandConfig, error) {
-	bandOpts := make([]PriorityBandConfigOption, 0, 3)
-	maxBytes, err := resolveQuantity(apiBand.MaxBytes, "DefaultPriorityBand MaxBytes")
-	if err != nil {
-		return nil, err
-	}
-	if maxBytes > 0 {
-		bandOpts = append(bandOpts, WithBandMaxBytes(maxBytes))
-	}
-	maxRequests, err := resolveQuantity(apiBand.MaxRequests, "DefaultPriorityBand MaxRequests")
-	if err != nil {
-		return nil, err
-	}
-	if maxRequests > 0 {
-		bandOpts = append(bandOpts, WithBandMaxRequests(maxRequests))
-	}
-	if apiBand.OrderingPolicyRef != "" {
-		bandOpts = append(bandOpts, WithOrderingPolicy(apiBand.OrderingPolicyRef, handle))
-	}
-	if apiBand.FairnessPolicyRef != "" {
-		bandOpts = append(bandOpts, WithFairnessPolicy(apiBand.FairnessPolicyRef, handle))
-	}
-
-	// We pass priority 0 as placeholder since it's a template.
-	templateBand, err := NewPriorityBandConfig(handle, 0, bandOpts...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create default priority band template: %w", err)
-	}
-	return templateBand, nil
-}
-
-func buildPriorityBand(handle plugin.Handle, band configapi.PriorityBandConfig) (*PriorityBandConfig, error) {
-	bandOpts := make([]PriorityBandConfigOption, 0, 3)
-	maxBytes, err := resolveQuantity(band.MaxBytes, fmt.Sprintf("priority band %d MaxBytes", band.Priority))
-	if err != nil {
-		return nil, err
-	}
-	if maxBytes > 0 {
-		bandOpts = append(bandOpts, WithBandMaxBytes(maxBytes))
-	}
-	maxRequests, err := resolveQuantity(band.MaxRequests, fmt.Sprintf("priority band %d MaxRequests", band.Priority))
-	if err != nil {
-		return nil, err
-	}
-	if maxRequests > 0 {
-		bandOpts = append(bandOpts, WithBandMaxRequests(maxRequests))
-	}
-	if band.OrderingPolicyRef != "" {
-		bandOpts = append(bandOpts, WithOrderingPolicy(band.OrderingPolicyRef, handle))
-	}
-	if band.FairnessPolicyRef != "" {
-		bandOpts = append(bandOpts, WithFairnessPolicy(band.FairnessPolicyRef, handle))
-	}
-
-	pb, err := NewPriorityBandConfig(handle, band.Priority, bandOpts...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create priority band config for priority %d: %w", band.Priority, err)
-	}
-	return pb, nil
-}
-
 // NewConfig creates a new Config populated with system defaults, applies the provided options, and enforces strict
 // validation.
 //
 // Arguments:
-//   - handle: A plugin.Handle required to resolve the default policies.
+//   - defaults: D`efault policy instances, provided by the config loader.
 //   - opts: Optional configuration overrides.
-func NewConfig(handle plugin.Handle, opts ...ConfigOption) (*Config, error) {
+func NewConfig(defaults PriorityBandPolicyDefaults, opts ...ConfigOption) (*Config, error) {
 	builder := &configBuilder{
 		config: &Config{
 			MaxBytes:              0, // no limit enforced
@@ -544,27 +388,27 @@ func NewConfig(handle plugin.Handle, opts ...ConfigOption) (*Config, error) {
 	// Initialize DefaultPriorityBand if missing.
 	// This ensures we always have a template for dynamic provisioning.
 	if builder.config.DefaultPriorityBand == nil {
-		template, err := NewPriorityBandConfig(handle, 0)
+		template, err := NewPriorityBandConfig(0, defaults)
 		if err != nil {
 			return nil, fmt.Errorf("failed to create default priority band: %w", err)
 		}
 		builder.config.DefaultPriorityBand = template
 	} else {
-		if err := builder.config.DefaultPriorityBand.applyDefaults(handle); err != nil {
+		if err := builder.config.DefaultPriorityBand.applyDefaults(defaults); err != nil {
 			return nil, fmt.Errorf("failed to apply defaults to DefaultPriorityBand: %w", err)
 		}
 	}
 
 	// Apply defaults to DefaultNegativePriorityBand if set.
 	if builder.config.DefaultNegativePriorityBand != nil {
-		if err := builder.config.DefaultNegativePriorityBand.applyDefaults(handle); err != nil {
+		if err := builder.config.DefaultNegativePriorityBand.applyDefaults(defaults); err != nil {
 			return nil, fmt.Errorf("failed to apply defaults to DefaultNegativePriorityBand: %w", err)
 		}
 	}
 
 	// Apply defaults to all explicitly configured bands.
 	for _, band := range builder.config.PriorityBands {
-		if err := band.applyDefaults(handle); err != nil {
+		if err := band.applyDefaults(defaults); err != nil {
 			return nil, fmt.Errorf("failed to apply defaults to priority band %d: %w", band.Priority, err)
 		}
 	}
@@ -578,8 +422,8 @@ func NewConfig(handle plugin.Handle, opts ...ConfigOption) (*Config, error) {
 // NewPriorityBandConfig creates a new band configuration with the required fields.
 // It applies system defaults first, then applies any provided options to override those defaults.
 func NewPriorityBandConfig(
-	handle plugin.Handle,
 	priority int,
+	defaults PriorityBandPolicyDefaults,
 	opts ...PriorityBandConfigOption,
 ) (*PriorityBandConfig, error) {
 	pb := &PriorityBandConfig{
@@ -592,7 +436,7 @@ func NewPriorityBandConfig(
 		}
 	}
 
-	if err := pb.applyDefaults(handle); err != nil {
+	if err := pb.applyDefaults(defaults); err != nil {
 		return nil, err
 	}
 
@@ -601,13 +445,12 @@ func NewPriorityBandConfig(
 
 // --- Validation, Defaults & Hydration ---
 
-func (p *PriorityBandConfig) applyDefaults(handle plugin.Handle) error {
+func (p *PriorityBandConfig) applyDefaults(defaults PriorityBandPolicyDefaults) error {
 	if p.OrderingPolicy == nil {
-		policy, err := orderingPolicy(DefaultOrderingPolicyRef, handle)
-		if err != nil {
-			return err
+		if defaults.OrderingPolicy == nil {
+			return fmt.Errorf("no default ordering policy provided and none set on band %d", p.Priority)
 		}
-		p.OrderingPolicy = policy
+		p.OrderingPolicy = defaults.OrderingPolicy
 	}
 	if p.Queue == "" {
 		p.Queue = defaultQueue
@@ -621,11 +464,10 @@ func (p *PriorityBandConfig) applyDefaults(handle plugin.Handle) error {
 		p.MaxBytes = defaultPriorityBandMaxBytes
 	}
 	if p.FairnessPolicy == nil {
-		policy, err := fairnessPolicy(DefaultFairnessPolicyRef, handle)
-		if err != nil {
-			return err
+		if defaults.FairnessPolicy == nil {
+			return fmt.Errorf("no default fairness policy provided and none set on band %d", p.Priority)
 		}
-		p.FairnessPolicy = policy
+		p.FairnessPolicy = defaults.FairnessPolicy
 	}
 	return nil
 }
